[ALLUXIO-1910] Move UfsUtils to cli package
