+++ shell/src/main/java/alluxio/cli/UfsUtils.java
@@ -9,12 +9,11 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.client;
+package alluxio.cli;
 
 import alluxio.AlluxioURI;
 import alluxio.Configuration;
 import alluxio.Constants;
-import alluxio.cli.Version;
 import alluxio.client.file.FileSystem;
 import alluxio.collections.PrefixList;
 import alluxio.exception.AlluxioException;
